third_party/blink/web_tests/external/wpt/webrtc/RTCDataChannelInit-maxRetransmits-enforce-range-expected.txt

This is a testharness.js-based test.
[FAIL] maxRetransmits with value -1 should throw TypeError
  assert_throws_js: function "() => {\n      pc.createDataChannel('test', { maxRetransmits: value });\n    }" did not throw
[FAIL] maxRetransmits with value -100 should throw TypeError
  assert_throws_js: function "() => {\n      pc.createDataChannel('test', { maxRetransmits: value });\n    }" did not throw
[FAIL] maxRetransmits with value 65536 should throw TypeError
  assert_throws_js: function "() => {\n      pc.createDataChannel('test', { maxRetransmits: value });\n    }" did not throw
[FAIL] maxRetransmits with value 100000 should throw TypeError
  assert_throws_js: function "() => {\n      pc.createDataChannel('test', { maxRetransmits: value });\n    }" did not throw
[FAIL] maxRetransmits with Infinity should throw TypeError
  assert_throws_js: function "() => {\n      pc.createDataChannel('test', { maxRetransmits: value });\n    }" did not throw
[FAIL] maxRetransmits with -Infinity should throw TypeError
  assert_throws_js: function "() => {\n      pc.createDataChannel('test', { maxRetransmits: value });\n    }" did not throw
[FAIL] maxRetransmits with NaN should throw TypeError
  assert_throws_js: function "() => {\n      pc.createDataChannel('test', { maxRetransmits: value });\n    }" did not throw
[FAIL] maxRetransmits with string "65536" should throw TypeError
  assert_throws_js: function "() => {\n      pc.createDataChannel('test', { maxRetransmits: value });\n    }" did not throw
Harness: the test ran to completion.
